move frontend to separate folder
